- fix login dan register - Mengganti email menjadi nama pada mainscreen
