+++ lib/ui/login_page.dart
@@ -1,6 +1,9 @@
-import 'package:firebase_auth/firebase_auth.dart';
+
 import 'package:flutter/material.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:fluttertoast/fluttertoast.dart';
 import 'package:petz_care/ui/main_screen.dart';
+
 import 'package:petz_care/ui/register_page.dart';
 
 class LoginPage extends StatefulWidget {
@@ -8,107 +11,202 @@ class LoginPage extends StatefulWidget {
   const LoginPage({Key? key}) : super(key: key);
 
   @override
-  State<LoginPage> createState() => _LoginPageState();
+  _LoginPageState createState() => _LoginPageState();
 }
 
 class _LoginPageState extends State<LoginPage> {
-  final _auth = FirebaseAuth.instance;
-  final _emailController = TextEditingController();
-  final _passwordController = TextEditingController();
+  final _formKey = GlobalKey<FormState>(); //form
 
-  bool _obscureText = true;
-  bool _isLoading = false;
+  final TextEditingController emailController =
+  new TextEditingController(); //container untuk form
+  final TextEditingController passwordController = new TextEditingController();
+
+  //firebase
+  final _auth = FirebaseAuth.instance;
+
+  bool isHiddenPassword = true;
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 24.0),
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: [
-            _isLoading
-                ? Center(child: CircularProgressIndicator())
-                : Container(),
-            Hero(
-              tag: 'Petz Care',
-              child: Text(
-                'Petz Care',
-                style: Theme.of(context).textTheme.headline5,
-              ),
-            ),
-            SizedBox(height: 24.0),
-            TextField(
-              controller: _emailController,
-              keyboardType: TextInputType.emailAddress,
-              decoration: InputDecoration(
-                border: OutlineInputBorder(),
-                hintText: 'Email',
-              ),
-            ),
-            SizedBox(height: 8.0),
-            TextField(
-              controller: _passwordController,
-              obscureText: _obscureText,
-              decoration: InputDecoration(
-                border: OutlineInputBorder(),
-                suffixIcon: IconButton(
-                  icon: Icon(
-                      _obscureText ? Icons.visibility : Icons.visibility_off),
-                  onPressed: () {
-                    setState(() {
-                      _obscureText = !_obscureText;
-                    });
-                  },
-                ),
-                hintText: 'Password',
-              ),
-            ),
-            SizedBox(height: 24.0),
-            MaterialButton(
-              child: Text('Login'),
-              color: Theme.of(context).primaryColor,
-              textTheme: ButtonTextTheme.primary,
-              height: 40,
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(16),
-              ),
-              onPressed: () async {
-                setState(() {
-                  _isLoading = true;
-                });
-                try {
-                  final email = _emailController.text;
-                  final password = _passwordController.text;
+    final emailField = TextFormField(
+      autofocus: false,
+      controller: emailController,
+      keyboardType: TextInputType.emailAddress,
+      validator: (value) {
+        if (value!.isEmpty) {
+          return ("Please Enter Your Email");
+        }
 
-                  await _auth.signInWithEmailAndPassword(
-                      email: email, password: password);
-                  Navigator.pushReplacementNamed(context, MainScreen.id);
-                } catch (e) {
-                  final snackbar = SnackBar(content: Text(e.toString()));
-                  ScaffoldMessenger.of(context).showSnackBar(snackbar);
-                } finally {
-                  setState(() {
-                    _isLoading = false;
-                  });
-                }
-              },
+        //req expression for email  validation
+        if (!RegExp("^[a-zA-Z0-9+_.-]+@[a-zA-Z0-9.-]+.[a-z]").hasMatch(value)) {
+          return ("Please Enter a valid email");
+        }
+        return null;
+      },
+      onSaved: (value) {
+        emailController.text = value!;
+      },
+      textInputAction: TextInputAction.next,
+      decoration: InputDecoration(
+        prefixIcon: Icon(Icons.mail),
+        contentPadding: EdgeInsets.fromLTRB(20, 15, 20, 15),
+        hintText: "Email",
+        border: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(10),
+        ),
+      ),
+    );
+
+    final passwordField = TextFormField(
+      controller: passwordController,
+      obscureText: isHiddenPassword,
+      validator: (value) {
+        RegExp regex = new RegExp(r'^.{6,}$'); //b
+        if (value!.isEmpty) {
+          return ("Password is required for login");
+        }
+        if (!regex.hasMatch(value)) {
+          return ("Enter Valid Password (Min.6Character)");
+        }
+      },
+      onSaved: (value) {
+        passwordController.text = value!;
+      },
+      textInputAction: TextInputAction.done,
+      decoration: InputDecoration(
+        prefixIcon: Icon(Icons.vpn_key),
+        contentPadding: EdgeInsets.fromLTRB(20, 15, 20, 15),
+        hintText: "Password",
+        border: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(10),
+        ),
+        suffixIcon: InkWell(
+          onTap: _togglePasswordView,
+          child: Icon(Icons.visibility),
+        ),
+      ),
+    );
+
+    final loginButton = Material(
+      elevation: 5,
+      borderRadius: BorderRadius.circular(30),
+      color: Colors.blueGrey,
+      child: MaterialButton(
+        padding: EdgeInsets.fromLTRB(20, 15, 20, 15),
+        minWidth: MediaQuery.of(context).size.width,
+        onPressed: () {
+          signIn(emailController.text, passwordController.text);
+        },
+        child: Text(
+          "Login",
+          textAlign: TextAlign.center,
+          style: TextStyle(
+            fontSize: 20,
+            color: Colors.white,
+            fontWeight: FontWeight.bold,
+          ),
+        ),
+      ),
+    );
+
+    return Scaffold(
+      backgroundColor: Colors.white,
+      body: Center(
+        child: SingleChildScrollView(
+          child: Container(
+            color: Colors.white10,
+            child: Padding(
+              padding: const EdgeInsets.all(36.0),
+              child: Form(
+                key: _formKey,
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  children: <Widget>[
+                    SizedBox(
+                      height: 50,
+                      child: Text(
+                        "Petz Care",
+                        style: TextStyle(
+                          fontSize: 24,
+                          fontWeight: FontWeight.bold,
+                          color: Colors.black,
+                          fontFamily: "Biryani",
+                        ),
+                      ),
+                    ),
+                    // SizedBox(
+                    //   width: 300,
+                    //   child: Image.asset(
+                    //     "assets/employee.jpg",
+                    //     fit: BoxFit.contain,
+                    //   ),
+                    // ),
+                    SizedBox(height: 45),
+                    emailField,
+                    SizedBox(height: 25),
+                    passwordField,
+                    SizedBox(height: 35),
+                    loginButton,
+                    SizedBox(height: 15),
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: <Widget>[
+                        Text("Don't have an account?"),
+                        GestureDetector(
+                          onTap: () {
+                            Navigator.push(
+                                context,
+                                MaterialPageRoute(
+                                    builder: (context) => RegistrationPage()));
+                          },
+                          child: Text(
+                            "SignUp",
+                            style: TextStyle(
+                              color: Colors.blueGrey,
+                              fontWeight: FontWeight.bold,
+                              fontSize: 15,
+                            ),
+                          ),
+                        ),
+                      ],
+                    ),
+                  ],
+                ),
+              ),
             ),
-            TextButton(
-              child: Text('Does not have an account yet? Register here'),
-              onPressed: () => Navigator.pushNamed(context, RegisterPage.id),
-            ),
-          ],
+          ),
         ),
       ),
     );
   }
 
-  @override
-  void dispose() {
-    _emailController.dispose();
-    _passwordController.dispose();
-    super.dispose();
+//login function
+
+  void signIn(String email, String password) async {
+    if (_formKey.currentState!.validate()) {
+      await _auth
+          .signInWithEmailAndPassword(email: email, password: password)
+          .then((uid) => {
+        Fluttertoast.showToast(msg: "Login Successful"),
+        Navigator.of(context).pushReplacement(
+            MaterialPageRoute(builder: (context) => MainScreen())),
+      })
+          .catchError((e) {
+        Fluttertoast.showToast(msg: e!.message);
+      });
+    }
+  }
+
+  //isHiddenPassword
+  void _togglePasswordView() {
+    // if (isHiddenPassword == true) {
+    //     isHiddenPassword = false;
+    // } else {
+    //   isHiddenPassword = true;
+    // }
+    setState(() {
+      isHiddenPassword = !isHiddenPassword;
+    });
   }
 }
