+++ lib/ui/register_page.dart
@@ -1,117 +1,336 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import 'package:fluttertoast/fluttertoast.dart';
+import 'package:petz_care/ui/main_screen.dart';
+import 'package:petz_care/ui/model_test.dart';
 
-class RegisterPage extends StatefulWidget {
-  static const String id = 'register_page';
-  const RegisterPage({Key? key}) : super(key: key);
+class RegistrationPage extends StatefulWidget {
+  static const String id = 'reg_page';
+
+  const RegistrationPage({Key? key}) : super(key: key);
 
   @override
-  _RegisterPageState createState() => _RegisterPageState();
+  _RegistrationPageState createState() => _RegistrationPageState();
 }
 
-class _RegisterPageState extends State<RegisterPage> {
-  final _auth = FirebaseAuth.instance;
-  final _emailController = TextEditingController();
-  final _passwordController = TextEditingController();
+class _RegistrationPageState extends State<RegistrationPage> {
+  final _formKey = GlobalKey<FormState>();
 
-  bool _obscureText = true;
-  bool _isLoading = false;
+  final _auth = FirebaseAuth.instance;
+
+  final TextEditingController firstNameController = new TextEditingController();
+  final TextEditingController secondNameController =
+      new TextEditingController();
+  final TextEditingController addressController = new TextEditingController();
+  final TextEditingController emailController = new TextEditingController();
+  final TextEditingController passwordController = new TextEditingController();
+  final TextEditingController confirmationpasswordController =
+      new TextEditingController();
+
+  bool isHiddenPassword = true;
+  bool isHiddenConfPassword = true;
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 24.0),
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: [
-            _isLoading
-                ? Center(child: CircularProgressIndicator())
-                : Container(),
-            Hero(
-              tag: 'Petz Care',
-              child: Text(
-                'Petz Care',
-                style: Theme.of(context).textTheme.headline5,
-              ),
-            ),
-            SizedBox(height: 24.0),
-            Text(
-              'Create your account',
-              style: Theme.of(context).textTheme.subtitle1,
-            ),
-            SizedBox(height: 8.0),
-            TextField(
-              controller: _emailController,
-              keyboardType: TextInputType.emailAddress,
-              decoration: InputDecoration(
-                border: OutlineInputBorder(),
-                hintText: 'Email',
-              ),
-            ),
-            SizedBox(height: 8.0),
-            TextField(
-              controller: _passwordController,
-              obscureText: _obscureText,
-              decoration: InputDecoration(
-                border: OutlineInputBorder(),
-                suffixIcon: IconButton(
-                  icon: Icon(
-                      _obscureText ? Icons.visibility : Icons.visibility_off),
-                  onPressed: () {
-                    setState(() {
-                      _obscureText = !_obscureText;
-                    });
-                  },
-                ),
-                hintText: 'Password',
-              ),
-            ),
-            SizedBox(height: 24.0),
-            MaterialButton(
-              child: Text('Register'),
-              color: Theme.of(context).primaryColor,
-              textTheme: ButtonTextTheme.primary,
-              height: 40,
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(16),
-              ),
-              onPressed: () async {
-                setState(() {
-                  _isLoading = true;
-                });
-                try {
-                  final email = _emailController.text;
-                  final password = _passwordController.text;
+    final firstNameField = TextFormField(
+      autofocus: false,
+      controller: firstNameController,
+      keyboardType: TextInputType.name,
+      validator: (value) {
+        RegExp regex = new RegExp(r'^.{3,}$');
+        if (value!.isEmpty) {
+          return ("First Name cannot be empty");
+        }
+        if (!regex.hasMatch(value)) {
+          return ("Enter Valid name (Min. 3 Character)");
+        }
+        return null;
+      },
+      onSaved: (value) {
+        firstNameController.text = value!;
+      },
+      textInputAction: TextInputAction.next,
+      decoration: InputDecoration(
+        prefixIcon: Icon(Icons.account_circle),
+        contentPadding: EdgeInsets.fromLTRB(20, 15, 20, 15),
+        hintText: "First Name",
+        border: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(10),
+        ),
+      ),
+    );
 
-                  await _auth.createUserWithEmailAndPassword(
-                      email: email, password: password);
-                  Navigator.pop(context);
-                } catch (e) {
-                  final snackbar = SnackBar(content: Text(e.toString()));
-                  ScaffoldMessenger.of(context).showSnackBar(snackbar);
-                } finally {
-                  setState(() {
-                    _isLoading = false;
-                  });
-                }
-              },
+    final secondNameField = TextFormField(
+      autofocus: false,
+      controller: secondNameController,
+      keyboardType: TextInputType.name,
+      validator: (value) {
+        // ignore: unused_local_variable
+        RegExp regex = new RegExp(r'^.{3,}$');
+        if (value!.isEmpty) {
+          return ("Second Name cannot be empty");
+        }
+        return null;
+      },
+      onSaved: (value) {
+        secondNameController.text = value!;
+      },
+      textInputAction: TextInputAction.next,
+      decoration: InputDecoration(
+        prefixIcon: Icon(Icons.account_circle),
+        contentPadding: EdgeInsets.fromLTRB(20, 15, 20, 15),
+        hintText: "Second Name",
+        border: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(10),
+        ),
+      ),
+    );
+
+    final addressField = TextFormField(
+      autofocus: false,
+      controller: addressController,
+      // keyboardType: TextInputType.number,
+      validator: (value) {
+        RegExp regex = new RegExp(r'^.{9,}$');
+        if (value!.isEmpty) {
+          return ("NIP is required for Register");
+        }
+        if (!regex.hasMatch(value)) {
+          return ("Enter Valid Address (Min.9 Character)");
+        }
+      },
+      onSaved: (value) {
+        addressController.text = value!;
+      },
+      textInputAction: TextInputAction.next,
+      decoration: InputDecoration(
+        prefixIcon: Icon(Icons.house),
+        contentPadding: EdgeInsets.fromLTRB(20, 15, 20, 15),
+        hintText: "Address",
+        border: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(10),
+        ),
+      ),
+    );
+
+    final emailField = TextFormField(
+      autofocus: false,
+      controller: emailController,
+      keyboardType: TextInputType.name,
+      validator: (value) {
+        if (value!.isEmpty) {
+          return ("Please Enter Your Email");
+        }
+
+        //req expression for email  validation
+        if (!RegExp("^[a-zA-Z0-9+_.-]+@[a-zA-Z0-9.-]+.[a-z]").hasMatch(value)) {
+          return ("Please Enter a valid email");
+        }
+        return null;
+      },
+      onSaved: (value) {
+        emailController.text = value!;
+      },
+      textInputAction: TextInputAction.next,
+      decoration: InputDecoration(
+        prefixIcon: Icon(Icons.mail),
+        contentPadding: EdgeInsets.fromLTRB(20, 15, 20, 15),
+        hintText: "Email",
+        border: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(10),
+        ),
+      ),
+    );
+
+    final passwordField = TextFormField(
+      autofocus: false,
+      controller: passwordController,
+      obscureText: isHiddenPassword,
+      validator: (value) {
+        RegExp regex = new RegExp(r'^.{6,}$');
+        if (value!.isEmpty) {
+          return ("Password is required for login");
+        }
+        if (!regex.hasMatch(value)) {
+          return ("Enter Valid Password (Min. 6 Character)");
+        }
+      },
+      onSaved: (value) {
+        passwordController.text = value!;
+      },
+      textInputAction: TextInputAction.done,
+      decoration: InputDecoration(
+        prefixIcon: Icon(Icons.vpn_key),
+        contentPadding: EdgeInsets.fromLTRB(20, 15, 20, 15),
+        hintText: "Password",
+        border: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(10),
+        ),
+        suffixIcon: InkWell(
+          onTap: _togglePasswordView,
+          child: Icon(Icons.visibility),
+        ),
+      ),
+    );
+
+    final confirmationpasswordField = TextFormField(
+      autofocus: false,
+      obscureText: isHiddenConfPassword,
+      controller: confirmationpasswordController,
+      validator: (value) {
+        if (confirmationpasswordController.text != passwordController.text) {
+          return "Password don't match ";
+        }
+        return null;
+      },
+      onSaved: (value) {
+        confirmationpasswordController.text = value!;
+      },
+      textInputAction: TextInputAction.next,
+      decoration: InputDecoration(
+        prefixIcon: Icon(Icons.vpn_key),
+        contentPadding: EdgeInsets.fromLTRB(20, 15, 20, 15),
+        hintText: "Confirm Password",
+        border: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(10),
+        ),
+        suffixIcon: InkWell(
+          onTap: _toggleConfPasswordView,
+          child: Icon(Icons.visibility),
+        ),
+      ),
+    );
+
+    final SignUpButton = Material(
+      elevation: 5,
+      borderRadius: BorderRadius.circular(30),
+      color: Colors.blueGrey,
+      child: MaterialButton(
+        padding: EdgeInsets.fromLTRB(20, 15, 20, 15),
+        minWidth: MediaQuery.of(context).size.width,
+        onPressed: () {
+          signUp(emailController.text, passwordController.text);
+        },
+        child: Text(
+          "Sign Up",
+          textAlign: TextAlign.center,
+          style: TextStyle(
+            fontSize: 20,
+            color: Colors.white,
+            fontWeight: FontWeight.bold,
+          ),
+        ),
+      ),
+    );
+
+    return Scaffold(
+      backgroundColor: Colors.white,
+      appBar: AppBar(
+        backgroundColor: Colors.transparent,
+        elevation: 0,
+        leading: IconButton(
+          icon: Icon(
+            Icons.arrow_back,
+            color: Colors.blueGrey,
+          ),
+          onPressed: () {
+            //kembali ke page login
+            Navigator.of(context).pop();
+          },
+        ),
+      ),
+      body: Center(
+        child: SingleChildScrollView(
+          child: Container(
+            color: Colors.white,
+            child: Padding(
+              padding: const EdgeInsets.all(20.0),
+              child: Form(
+                key: _formKey,
+                child: Column(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    crossAxisAlignment: CrossAxisAlignment.center,
+                    children: <Widget>[
+                      // SizedBox(
+                      //     width: 200,
+                      //     child: Image.asset(
+                      //       "assets/employee.jpg",
+                      //       fit: BoxFit.contain,
+                      //     )),
+                      SizedBox(height: 20),
+                      firstNameField,
+                      SizedBox(height: 20),
+                      secondNameField,
+                      SizedBox(height: 20),
+                      addressField,
+                      SizedBox(height: 20),
+                      emailField,
+                      SizedBox(height: 20),
+                      passwordField,
+                      SizedBox(height: 20),
+                      confirmationpasswordField,
+                      SizedBox(height: 20),
+                      SignUpButton,
+                      SizedBox(height: 15),
+                    ]),
+              ),
             ),
-            TextButton(
-              child: Text('Already have an account? Login'),
-              onPressed: () => Navigator.pop(context),
-            ),
-          ],
+          ),
         ),
       ),
     );
   }
 
-  @override
-  void dispose() {
-    _emailController.dispose();
-    _passwordController.dispose();
-    super.dispose();
+  void signUp(String email, String password) async {
+    if (_formKey.currentState!.validate()) {
+      await _auth
+          .createUserWithEmailAndPassword(email: email, password: password)
+          .then((value) => {postDetailToFirestore()})
+          .catchError((e) {
+        Fluttertoast.showToast(msg: e!.message);
+      });
+    }
+  }
+
+  postDetailToFirestore() async {
+    //calling our firestore
+    //calling our user model
+    //sedning these values
+
+    FirebaseFirestore firebaseFirestore = FirebaseFirestore.instance;
+    User? user = _auth.currentUser;
+    UserModel userModel = UserModel();
+
+    // writing all the values
+    userModel.email = user!.email;
+    userModel.uid = user.uid;
+    userModel.firstName = firstNameController.text;
+    userModel.secondName = secondNameController.text;
+    userModel.address = addressController.text;
+
+    await firebaseFirestore
+        .collection("user")
+        .doc(user.uid)
+        .set(userModel.toMap());
+
+    Fluttertoast.showToast(msg: "Account created successfully :)");
+
+    Navigator.pushAndRemoveUntil(
+        context,
+        MaterialPageRoute(builder: (context) => MainScreen()),
+        (route) => false);
+  }
+
+  void _togglePasswordView() {
+    setState(() {
+      isHiddenPassword = !isHiddenPassword;
+    });
+  }
+
+  void _toggleConfPasswordView() {
+    setState(() {
+      isHiddenConfPassword = !isHiddenConfPassword;
+    });
   }
 }
